Revert "Merge branch 'main' into onsen-allocator"
This reverts commit 481cf247d49464dfbdd713c7b965d2b40cfbd862, reversing
changes made to 23c96d20c70139a930d40d0ac048e0caf3c4c27b.
