+++ contracts/interfaces/IYieldDirector.sol
@@ -2,17 +2,9 @@
 pragma solidity >=0.7.5;
 
 interface IYieldDirector {
-    function deposit(uint256 amount_, address recipient_) external returns (uint256);
+    function deposit(uint256 amount_, address recipient_) external;
 
-    function depositSohm(uint256 amount_, address recipient_) external returns (uint256);
-
-    function addToDeposit(uint256 depositId_, uint256 amount_) external;
-
-    function addToSohmDeposit(uint256 depositId_, uint256 amount_) external;
-
-    function withdrawPrincipal(uint256 depositId, uint256 amount_) external;
-
-    function withdrawPrincipalAsSohm(uint256 depositId_, uint256 amount_) external;
+    function withdraw(uint256 amount_, address recipient_) external;
 
     function withdrawAll() external;
 
@@ -26,19 +18,7 @@ interface IYieldDirector {
 
     function totalDonated(address donor_) external view returns (uint256);
 
-    function getDepositorIds(address donor_) external view returns (uint256[] memory);
+    function redeem() external;
 
-    function redeemYield(uint256 depositId_) external;
-
-    function redeemYieldAsSohm(uint256 depositId_) external;
-
-    function redeemAllYield() external;
-
-    function redeemAllYieldAsSohm() external;
-
-    function redeemableBalance(uint256 depositId_) external view returns (uint256);
-
-    function totalRedeemableBalance(address recipient_) external view returns (uint256);
-
-    function getRecipientIds(address recipient_) external view returns (uint256[] memory);
+    function redeemableBalance(address recipient_) external view returns (uint256);
 }
