+++ contracts/vesting/Investor.sol
@@ -12,7 +12,7 @@ import "../types/Ownable.sol";
 interface IClaim {
     struct Term {
         uint256 percent; // 4 decimals ( 5000 = 0.5% )
-        uint256 gClaimed; // static number
+        uint256 wClaimed; // static number
         uint256 max; // maximum nominal OHM amount can claim
     }
 
@@ -57,7 +57,7 @@ contract InvestorClaimV2 is Ownable {
     // tracks rebase-agnostic balance
     IgOHM internal immutable gOHM = IgOHM(0x0ab87046fBb341D058F17CBC4c1133F25a20a52f);
     // previous deployment of contract (to migrate terms)
-    IClaim internal immutable previous = IClaim(0xcD4B3c7B746161f0E54bc9a23307CE222a2bF081);
+    IClaim internal immutable previous = IClaim(0xaCCC8306455BaA01593Fa6267809fEA72F684169);
 
     // tracks address info
     mapping(address => Term) public terms;
@@ -129,7 +129,7 @@ contract InvestorClaimV2 is Ownable {
      */
     function pullWalletChange(address _oldAddress) external {
         require(walletChange[_oldAddress] == msg.sender, "Old wallet did not push");
-        require(terms[msg.sender].percent == 0, "Wallet already exists");
+        require(terms[msg.sender].percent != 0, "Wallet already exists");
 
         walletChange[_oldAddress] = address(0);
         terms[msg.sender] = terms[_oldAddress];
@@ -185,7 +185,7 @@ contract InvestorClaimV2 is Ownable {
     function migrate(address[] memory _addresses) external onlyOwner {
         for (uint256 i = 0; i < _addresses.length; i++) {
             IClaim.Term memory term = previous.terms(_addresses[i]);
-            setTerms(_addresses[i], term.percent, term.gClaimed, term.max);
+            setTerms(_addresses[i], term.percent, term.wClaimed, term.max);
         }
     }
