+++ docs/deployments/ethereum.md
@@ -1,32 +1,21 @@
 # 📜 Ethereum Addresses
 
 ## Core
-
+---
 | Contract | Address |
 | --- | --- |
 OHM V2 | `0x64aa3364F17a4D01c6f1751Fd97C2BD3D7e7f1D5`
 sOHM V2 | `0x04906695D6D12CF5459975d7C3C03356E4Ccd460`
 gOHM | `0x0ab87046fBb341D058F17CBC4c1133F25a20a52f`
 Treasury V2 | `0x9A315BdF513367C0377FB36545857d12e85813Ef`
-TreasuryExtender | `0xb32Ad041f23eAfd682F57fCe31d3eA4fd92D17af`
 Staking | `0xB63cac384247597756545b500253ff8E607a8020`
 Distributor | `0xeeeb97A127a342656191E0313DF33D58D06B2E05`
 BondDepositoryV2 | `0x9025046c6fb25Fb39e720d97a8FD881ED69a1Ef6`
 Authority | `0x1c21F8EA7e39E2BA00BC12d2968D63F4acb38b7A`
 
 ## Multisigs
-
+---
 | Contract | Address |
 | --- | --- |
 DAO | `0x245cc372C84B3645Bf0Ffe6538620B04a217988B`
 Policy | `0x0cf30dc0d48604A301dF8010cdc028C055336b2E`
-
-## Allocators
-
-| Contract | Address |
-| --- | --- |
-AaveAllocatorV2 | `0x0D33c811D0fcC711BcB388DFB3a152DE445bE66F`
-AlchemixAllocatorV2 | N/A in fix
-FxsAllocatorV2 | `0x0f953D861347414698F34B75dbFd6e7dF1A73493`
-BtrflyAllocator | `0xC8431fEb345B46c30A4576c1b5faF080fdc54e2f`
-LUSDAllocatorV2 | `0x2C1700F38C38C32595CFeF3D6B0B275bC2D2a578`
